@@ -4,7 +4,7 @@
   "main": "expo-router/entry",
   "scripts": {
     "start": "expo start",
-    "test:dev": "jest --watchAll",
+    "test:dev": "jest --watchAll --passWithNoTests",
     "test": "jest",
     "android": "expo run:android",
     "ios": "expo run:ios",
@@ -17,40 +17,42 @@
     "@expo-google-fonts/shrikhand": "^0.2.3",
     "@react-navigation/native": "^7.0.0",
     "@testing-library/jest-dom": "^6.6.3",
-    "@types/react": "~18.3.12",
+    "@types/react": "~19.1.10",
     "axios": "^1.6.7",
     "babel-plugin-syntax-hermes-parser": "^0.26.0",
-    "expo": "~52.0.36",
-    "expo-auth-session": "~6.0.3",
-    "expo-cellular": "~7.0.2",
-    "expo-checkbox": "~4.0.1",
-    "expo-constants": "~17.0.6",
-    "expo-crypto": "~14.0.2",
-    "expo-font": "~13.0.3",
-    "expo-image": "~2.0.6",
-    "expo-linear-gradient": "~14.0.2",
-    "expo-linking": "~7.0.5",
-    "expo-location": "~18.0.7",
-    "expo-router": "~4.0.17",
-    "expo-secure-store": "~14.0.1",
-    "expo-status-bar": "~2.0.1",
-    "expo-web-browser": "~14.0.2",
+    "expo": "^54.0.2",
+    "expo-auth-session": "~7.0.8",
+    "expo-cellular": "~8.0.7",
+    "expo-checkbox": "~5.0.7",
+    "expo-constants": "~18.0.8",
+    "expo-crypto": "~15.0.7",
+    "expo-font": "~14.0.8",
+    "expo-image": "~3.0.8",
+    "expo-linear-gradient": "~15.0.7",
+    "expo-linking": "~8.0.8",
+    "expo-location": "~19.0.7",
+    "expo-maps": "^0.12.7",
+    "expo-router": "~6.0.1",
+    "expo-secure-store": "~15.0.7",
+    "expo-status-bar": "~3.0.8",
+    "expo-web-browser": "~15.0.7",
     "jest": "^29.7.0",
-    "jest-expo": "~52.0.5",
-    "react": "^18.2.0",
+    "jest-expo": "~54.0.10",
+    "react": "19.1.0",
     "react-devtools-core": "^5.3.1",
-    "react-dom": "^18.2.0",
-    "react-native": "^0.73.9",
-    "react-native-maps": "1.18.0",
+    "react-dom": "19.1.0",
+    "react-native": "0.81.4",
+    "react-native-maps": "1.20.1",
     "react-native-maps-directions": "^1.9.0",
-    "react-native-reanimated": "~3.16.1",
-    "react-native-safe-area-context": "4.12.0",
-    "react-native-screens": "~4.4.0",
-    "react-native-svg": "15.8.0",
+    "react-native-reanimated": "~4.1.0",
+    "react-native-safe-area-context": "~5.6.0",
+    "react-native-screens": "~4.16.0",
+    "react-native-svg": "15.12.1",
     "react-native-vector-icons": "^10.0.3",
-    "react-test-renderer": "18.2.0",
-    "typescript": "^5.3.0",
-    "zod": "^3.24.2"
+    "react-test-renderer": "^19.1.1",
+    "typescript": "~5.9.2",
+    "zod": "^3.24.2",
+    "react-native-worklets": "0.5.1"
   },
   "devDependencies": {
     "@babel/core": "^7.23.0",
@@ -74,5 +76,10 @@
     "react-native-svg-transformer": "^1.3.0",
     "typescript-eslint": "^8.15.0"
   },
-  "private": true
+  "private": true,
+  "overrides": {
+    "react-test-renderer": {
+      "react": "^19.1.0"
+    }
+  }
 }
