@@ -1,9 +1,9 @@
-module.exports = function (api) {
+export default function (api) {
   api.cache(true);
   return {
     presets: [["babel-preset-expo"], "@babel/preset-typescript"],
     plugins: [
-      'react-native-reanimated/plugin',
+      'react-native-worklets/plugin',
     ]
   };
-};
+}
